+++ entry/api_v2/urls.py
@@ -58,4 +58,5 @@ urlpatterns = [
         ),
     ),
     path("advanced_search/", views.AdvancedSearchAPI.as_view()),
+    path("import/", views.EntryImportAPI.as_view()),
 ]