+++ frontend/src/apiclient/autogenerated/apis/index.ts
@@ -4,4 +4,5 @@ export * from "./AclApi";
 export * from "./EntityApi";
 export * from "./EntryApi";
 export * from "./GroupApi";
+export * from "./JobApi";
 export * from "./UserApi";